Issue #42: Added support for narrow/widen in ztree-dir.
In ztree-dir now it is possible to focus on directory under the cursor
(or parent directory of the file under the cursor) using '>' hotkey.

One can also widen to the to the parent of the root directory
using '<' hotkey.

This allows to navigate the whole filesystem without leaving the buffer.