Issue #61: Error on unaccessible directory contents
When ztree-show-number-of-children is on, ztree-dir calculates
the amount of children in each subdirectory of the current
directory to present the number of elements in square braces
[].
If the directory is inaccessible due to lack of rights it lead
to error.
Now the error is ignored and "N/A" text shown instead of number
of elements in a directory.